+++ tests/receivers/lightprometheus/internal_prometheus_test.go
@@ -18,14 +18,23 @@
 package tests
 
 import (
+	"path"
 	"testing"
 
 	"github.com/signalfx/splunk-otel-collector/tests/testutils"
 )
+
+var httpd = []testutils.Container{testutils.NewContainer().WithContext(
+	path.Join(".", "testdata", "server"),
+).WithName("httpd").WithExposedPorts("8000:80").WillWaitForLogs("httpd -D FOREGROUND")}
 
 func TestInternalPrometheusMetrics(t *testing.T) {
 	testutils.SkipIfNotContainerTest(t) // TODO: enhance internal metric settings detection for process config
 	testutils.AssertAllMetricsReceived(
 		t, "internal.yaml", "internal_metrics_config.yaml", nil, nil,
 	)
+}
+
+func TestHttpdBasicAuth(t *testing.T) {
+	testutils.AssertAllMetricsReceived(t, "basic_auth_metrics.yaml", "httpd_basic_auth.yaml", httpd, nil)
 }